// src/types/tiktok.ts
export interface TikTokAccountInfo {
  username: string;
  nickname: string;
  avatar: string;
  bio: string;
  countryName: string;
  countryCode: string;
  flag: string;
  profileUrl: string;
  followers: string;
  following: string;
  likes: string;
  videos: string;
}

export interface TikTokError {
  message: string;
}